finishing dividing the components and theres a typescript problem remaining
